requirements.txt
streamlit>=1.32.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.119.0
notion-client>=2.2.1
langchain>=0.1.10
langchain-openai>=0.0.8
python-dotenv>=1.0.1
requests>=2.31.0
Pillow>=10.2.0
pytz>=2024.1
python-dateutil>=2.8.2
